Released 1.2.10.
Sun Apr  1 23:01:31 2001  Owen Taylor  <otaylor@redhat.com>

	* Released 1.2.10.

Sun Apr  1 22:22:47 2001  Owen Taylor  <otaylor@redhat.com>

	[ Addition of safety checks to try to avoid segfaulting
	  on broken setups always. ]

	* gdk/gdkfont.c (_gdk_font_wc_to_glyphs): Add a boolean
	return value to deal with failure of conversion of
	wide characters to glyphs.

	* gtk/gtkentry.c gtk/gtklabel.c: Handle failure of
	mbstowcs/wcstombs a bit better.

	* gdk/gdkfont.c gdk/gdkdraw.c: Deal with failure of
	_gdk_font_wc_to_glyphs() gracefully

	* gdk/gdkim.c (_gdk_wcstombs_len): quietly return NULL on
	failure instead of g_return_if_fail().

Wed Mar 28 16:05:29 2001  Owen Taylor  <otaylor@redhat.com>

	* gtk/gtkfontsel.c (gtk_font_selection_select_size): Fix
	problem where fractional sizes caused infinite loop.
	(bugzilla.redhat.com #33081)

Wed Mar 28 14:52:08 2001  Owen Taylor  <otaylor@redhat.com>

	* docs/Makefile.am: Remove excess rules breaking
	distcheck.

+++ gtk/gtkfontsel.c
@@ -1647,7 +1647,7 @@ gtk_font_selection_select_size (GtkWidget      *w,
 				gpointer        data)
 {
   GtkFontSelection *fontsel;
-  gint new_size;
+  gdouble new_size;
   gchar *text;
   gchar buffer[16];
   gint i;
@@ -1673,11 +1673,11 @@ gtk_font_selection_select_size (GtkWidget      *w,
   gtk_entry_set_text(GTK_ENTRY(fontsel->size_entry), buffer);
   
   /* Check if the font size has changed, and return if it hasn't. */
-  new_size = atoi(text);
+  new_size = atof(text);
   if (fontsel->metric == GTK_FONT_METRIC_POINTS)
     new_size *= 10;
   
-  if (fontsel->size == new_size)
+  if (fontsel->size == (gint)new_size)
     return;
   
   /* If the size was selected by the user we set the selected_size. */
